chore(e2e-suite): archive sel btc fee test that was disabled for 6 months
